remove statement cache
prepareCached is not that useful and implementation is somewhat ugly. It
is the task of the application to keep prepared statements when used multiple
times.

+++ test/types-test.cpp
@@ -454,8 +454,8 @@ public:
         tester.testFloat(static_cast<float>(-123));
         tester.testFloat(std::numeric_limits<float>::max() * .999999);
         tester.testFloat(std::numeric_limits<float>::min() * 1.00001);
-        tester.test(std::numeric_limits<double>::infinity());
-        tester.test(-std::numeric_limits<double>::infinity());
+        tester.test(std::numeric_limits<float>::infinity());
+        tester.test(-std::numeric_limits<float>::infinity());
     }
 
     void testFloatFromDecimal()
